do not load minitest/ci unless required
Minitest.load_plugins will automatically load all plugins including
'minitest/ci_plugin'. Before, this would not only load the gem but
also adds a new Minitest::Ci object to Minitest.reporters as long as
the gem exists anywhere in $LOAD_PATH.

This change requires that 'minitest/ci' be required explicitly or be
implicitly required by Bundler to be added as a reporter.